% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-df.R
\name{study_df}
\alias{study_df}
\title{Cast Simulation Study Matrices to Data.Frame}
\usage{
study_df(..., wide = TRUE, data_names = NULL)
}
\arguments{
\item{...}{A list of `matrices`}

\item{wide}{A `boolean` indicating if the simulated data is group
by row (`TRUE`) or column (`FALSE`).}

\item{data_names}{A `character vector` containing name of matrix. If
empty it uses the simulation matrix name}
}
\value{
A `data.frame` with three variables:

- `Round`: Iteration of the Simulation
- `Draw`: Draw during the iteration of the simulation
- `Value`: Value of the statistic at round and draw
- `Type`: Study Matrix
}
\description{
Creates a `data.frame` out of multiple simulation matrices
for use with `ggplot2`.
}
\examples{
# Set Seed
set.seed(5812)

# Generate data
m1 = matrix(rnorm(10), 2, 5)
m2 = matrix(rnorm(10), 2, 5)
m3 = matrix(rnorm(10), 2, 5)

# Organize data.frame by row
study_df(m1, m2, m3)

# Organize by column
study_df(m1, m2, m3, wide = FALSE,
         data_names = c("Hello", "Goodbye", "Wabbit"))
}
\author{
James Joseph Balamuta
}
