% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysim-package.R
\docType{package}
\name{tidysim-package}
\alias{tidysim}
\alias{tidysim-package}
\title{tidysim: Tidy Simulation}
\description{
Provides functions for performing simulations that generate 
    tidy data sets.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item David Dalpiaz \email{dalpiaz2@illinois.edu} [copyright holder]
}

}
\keyword{internal}
