% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-df.R
\name{sim_df}
\alias{sim_df}
\title{Cast Simulation Matrix to Data Frame}
\usage{
sim_df(m, wide = TRUE)
}
\arguments{
\item{m}{A `matrix`}

\item{wide}{A `boolean` indicating if the simulated data is group
by row (`TRUE`) or column (`FALSE`). Default: By row (`TRUE`).}
}
\value{
A `data.frame` with three variables:

- `Round`: Iteration of the Simulation
- `Parameter`: Parameter that was Estimated during Simulation
- `Value`: Value of the statistic at round and draw
}
\description{
Creates a `data.frame` out of a simulation matrix
for use with `ggplot2`.
}
\examples{
# Set Seed
set.seed(5812)

# Generate data
m = matrix(rnorm(10), 2,5)

# Organize data by column, e.g.
# Each column is a repeated measurement
# Each row is a variable
sim_df(m)

# Organize the data by row, e.g.
# Each row is a repeated measurement
# Each column is a variable
sim_df(m, wide = FALSE)
}
\author{
James Joseph Balamuta
}
